requirements.txt
moviepy==1.0.3
protobuf==5.28.2
pydub==0.25.1
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.39.0
google-cloud-storage==2.18.2
google-cloud-speech==2.27.0  
google-cloud-texttospeech==2.18.0 
google-auth== 2.35.0             